requirements.txt
boto3~=1.34.103
botocore~=1.34.103
python-dotenv~=1.0.1
streamlit~=1.34.0
cohere~=5.1.7